requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.20.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
redis==4.3.3
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
